requirements.txt
ciso8601==2.3.2
influxdb_client==1.48.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
reactivex==4.0.4
typing_extensions==4.12.2
numpy==1.24.3
scipy==1.10.1